back/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
Bottleneck==1.3.7
Brotli==1.0.9
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==2.0.4
cryptography==42.0.5
Django==5.0.4
django-cors-headers==4.3.1
django-environ==0.11.2
django-nextjs==2.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numexpr==2.8.7
numpy==1.26.4
pandas==2.2.1
pip==23.3.1
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2023.3
urllib3==1.26.18
websockets==12.0
wheel==0.41.2
yarl==1.9.4
